KEXP: Able to plot normalised abundances + time as xvar in colour plot
Several options are available for both features. The normalisation value 
can be choosen but the default is the first species at time zero. When 
using time as xvar, the origianl xvar (e.g. density) has to be fixed. 
This can be choosen with xvar_point. Previous options such as same 
colour scale for all plots also work with the new features.